messy

+++ udp.h
@@ -46,11 +46,11 @@ typedef struct sockaddr_in SAI;
 
 int Socket( int family, int type, int protocol);
 
-int senddata( int fd, void *data, size_t size,
+int senddata( int fd, void *data, int size,
              const struct sockaddr *servaddr, socklen_t addrlen,
              int logged);
 
-int Read( int sockfd, void *data, size_t size, int logged);
+int Read( int sockfd, void *data, int size, int logged);
 
 int clientinit( char* hostname, char* port, SAI** sock_addr);
